requirements.txt
flask
numpy
pillow
opencv-python-headless